+++ TagsCloudVisualization/WordsPrepares/WordsPreparer.cs
@@ -1,7 +1,7 @@
 ﻿using System.Collections.Generic;
 using System.Linq;
 
-namespace TagsCloudVisualization.WordsPrepare
+namespace TagsCloudVisualization.WordsPrepares
 {
     public class WordsPreparer : IWordsPreparer
     {
