+++ TagsCloudVisualization/WordsToTagTransformers/IWordsToTagTransformer.cs
@@ -1,6 +1,6 @@
 ﻿using System.Collections.Generic;
 
-namespace TagsCloudVisualization.WordsToTagTransformer
+namespace TagsCloudVisualization.WordsToTagTransformers
 {
     public interface IWordsToTagTransformer
     {
